Improved extensibility
